+++ R/derive_merged.R
@@ -406,7 +406,7 @@ derive_vars_merged <- function(dataset,
   assert_data_frame(dataset, required_vars = by_vars_left)
   assert_data_frame(
     dataset_add,
-    required_vars = expr_c(
+    required_vars = c(
       by_vars_right,
       setdiff(extract_vars(order), replace_values_by_names(new_vars)),
       extract_vars(new_vars)
@@ -1139,7 +1139,7 @@ derive_vars_merged_summary <- function(dataset,
   )
   assert_data_frame(
     dataset_add,
-    required_vars = expr_c(by_vars_right, extract_vars(new_vars))
+    required_vars = c(by_vars_right, extract_vars(new_vars))
   )
 
   # Summarise the analysis value and merge to the original dataset