+++ NEWS.md
@@ -5,8 +5,12 @@
 ## Updates of Existing Functions
 
 - The `min_dates_strict` and `max_dates_strict` arguments were added to
-  `derive_vars_dt()` and `derive_vars_dtm()` and other imputation functions. See
-  `derive_vars_dt()` documentation for details and examples. (#2708) 
+`derive_vars_dt()`, `derive_vars_dtm()`, `convert_dtc_to_dt()`,
+`convert_dtc_to_dtm()`, `impute_dtc_dt()`, `impute_dtc_dtm()`. The new arguments
+can be used to avoid invalid imputed dates. E.g., if `AESTDTC` is imputed with
+`min_dates = exprs(TRTSDTM)`, `max_dates_strict = exprs(AENDTM)` can be
+specified to avoid that the imputed event start date is after the even end date.
+See `derive_vars_dt()` documentation for details and examples. (#2708)
 
 ## Breaking Changes